Implemented status, moved cooldown to cog manager
