+++ cogs/dice.py
@@ -1,35 +1,15 @@
 from discord import Embed, Colour, app_commands, Interaction
 from discord.ext import commands
 import random as rand
-import time
+from cogs.cog_manager import command_cooldown
 from datetime import datetime
 
 dice_options = [
     "d20", "d100", "d4", "d6", "d10", "d12", "d10000", "d1000000"
 ]
 dice_choices = [app_commands.Choice(name=dice, value=dice) for dice in dice_options]
-dice_cog_cooldowns = {}
 SPECIAL_NUMBERS = {69, 420, 42069, 69420, 999999}
 
-def dice_cog_cooldown(seconds: int):
-    async def predicate(interaction: Interaction):
-        user_id = interaction.user.id
-        current_time = time.time()
-        
-        if user_id in dice_cog_cooldowns:
-            last_used = dice_cog_cooldowns[user_id]
-            if current_time - last_used < seconds:
-                cooldown_remaining = seconds - (current_time - last_used)
-                await interaction.response.send_message(
-                    f"You're on cooldown! Try again in {cooldown_remaining:.1f} seconds.",
-                    ephemeral=True
-                )
-                return False
-
-        dice_cog_cooldowns[user_id] = current_time
-        return True
-
-    return app_commands.check(predicate)
 
 class Dice(commands.Cog, name="Dice"):
     def __init__(self, bot) -> None:
@@ -48,7 +28,7 @@ class Dice(commands.Cog, name="Dice"):
     @app_commands.command(name="coin", description="Flip a coin!")
     @app_commands.describe(what_for="What are you flipping the coin for?")
     @app_commands.rename(what_for="for")
-    @dice_cog_cooldown(seconds=5)
+    @command_cooldown(seconds=5)
     async def coin_flip(self, interaction: Interaction, what_for: str = ''):
         try:
             coin = rand.randint(0,999)
@@ -79,7 +59,7 @@ class Dice(commands.Cog, name="Dice"):
     @app_commands.describe(what_for="What are you rolling for?")
     @app_commands.rename(what_for="for")
     @app_commands.choices(dice=dice_choices)
-    @dice_cog_cooldown(seconds=5)
+    @command_cooldown(seconds=5)
     async def roll_number(self, interaction: Interaction, dice: str, what_for: str = ''):
         max_roll: int = 0
         if dice.startswith("d"):
@@ -112,7 +92,7 @@ class Dice(commands.Cog, name="Dice"):
     @app_commands.describe(what_for="What are you rolling for?")
     @app_commands.choices(dice=dice_choices)
     @app_commands.rename(what_for="for")
-    @dice_cog_cooldown(seconds=5)
+    @command_cooldown(seconds=5)
     async def advantage(self, interaction: Interaction, dice: str, what_for: str = ''):
         max_roll: int = 0
         if dice.startswith("d"):
@@ -147,10 +127,6 @@ class Dice(commands.Cog, name="Dice"):
             await interaction.response.send("Oops! Something went wrong.", ephemeral=True)
             print(e)
         
-        
-
-
-    # future implementation of adding special numbers
 
 async def setup(client):
     await client.add_cog(Dice(client))
